Testing against provider.data service
The service uses form encoding for its data.  The template stuff, etc
worked well but realised I needed to set the correct content type.  The
content type was hard coded so have setup in the config to allow a set
of headers to be set in a map.

The yaml is of the form

headers:
    h1: [value1]
    h2: [value2]

I've used the http.Header for the struct in config - this ensures its
usable in the headers of the http request.