Changed to use the filename as results URL
If the file exists charthandler will still load the data from it.
Either way it now also uses the filename as the part of the URL of the
results

Adjusted tests to cover this as well.

Also found an interesting way of testing something that hits a
log.Fatal.  You can spin the test up in a separate process and then
check for the error code returned - nice.
See http://stackoverflow.com/a/34247912 for details